requirements.txt
Flask==1.1.1
uWSGI==2.0.18
redis==3.3.11
Pyrebase==3.0.27
Pillow==6.2.1
opencv-python==4.1.2.30
pandas==0.25.3
scikit-learn==0.21.3
fastai==1.0.59